moved and renamed LocalizationProvider file
